+++ test/libsolidity/syntaxTests/comments/multiline_unicode_isolate_override_2.sol
@@ -1,0 +1,14 @@
+contract C {
+    function f() public pure
+    {
+        // LRI PDI
+        /*ok ⁦⁩*/
+
+        // LRI LRI PDI PDI
+        /*ok ⁦⁦⁩⁩*/
+
+        // RLI FSI PDI PDI
+        /*ok ⁧⁨⁩⁩*/
+    }
+}
+// ----